requirements.txt

streamlit==1.42.0
pymongo==4.11.1
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.1
scipy==1.15.1
tqdm==4.67.1
python-docx==1.1.2


sentence-transformers==3.4.1
transformers==4.48.3
huggingface-hub==0.28.1
langchain==0.3.18
openai==1.63.0
tiktoken==0.9.0

pdf2image==1.17.0
PyMuPDF==1.25.3
pytesseract==0.3.13
Pillow==11.1.0

PyYAML==6.0.2
filelock==3.17.0
tenacity==9.0.0
